requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
openai>=1.30.0
asyncpg
python-multipart
openai-agents
pypdf
